new ForwardDeclarations.h
